rvCAN/can-test/requirements.txt
#pillow
#numpy
#RPi.GPIO
#spidev 
python-can